A lot of changes made by the Eigenbase/LucidDB project.
This version of sqlline was forked from
sqlline.sourceforge.net's CVS repository after 5-June-2006 for
easier maintenance within The Eigenbase Project.

These are some of the changes which were made to the Eigenbase
version in between when the fork occurred and when the unarchived source
code was checked into perforce.eigenbase.org under //open/sqlline:

- Treat multiline statements as being on multiple lines instead
  of being all on the same line (Angel)

- Add --run command-line option (John Pham and Sunny Choi)

- Make --showWarnings default to true and take effect even when
  verbose is off; show warnings immediately after connect (JVS)

- Fix ColorBuffer performance problem with very long rows (JVS)

- Use driver.connect() rather than DriverManager.getConnection()
  to avoid deadlocks with drivers like RMI-JDBC (Steve Herskovitz)

- In signal handler null out stmt after handling to avoid problems with
  double crtl-c press. (Jack Hahn)

- Standardized indentation and ran Jalopy; still need to break
  up SqlLine.java (JVS)

Descriptions of subsequent changes can be found in Perforce.

For more information, see

http://pub.eigenbase.org/wiki/SqllineDocs