[SQLLINE-287] Support JDK 12
Document JDK 12 as our preferred JDK version; anything as old as JDK 8
is fine.

Use JDK 12 to build releases, and add JDK 12 to the list of JDKs to
test each release on.

Enable Travis on JDK 12. (Also enable Travis on JDK 13, which is still
under development.)

Enable AppVeyor on JDK 12.