[SQLLINE-331] Autocompletion for schema, table, column names
Data for autocompletion is loaded by steps (columns names are loaded
after the first call of completion for columns of this table).

SQL identifier quotes are automatically added if there are spaces,
escaped identifiers, case-sensitive or user already started with SQL
identifier.

Add respect of getExtraNameCharacters while autocompletion with
autoquotation.